fixes, add tests
